+++ tests/unit/converters/schema/test_halogen.py
@@ -1,9 +1,18 @@
 from http import HTTPStatus
 
+import halogen
 import pytest
+from flask import jsonify
 
-from openapi_builder import set_schema_options
+from openapi_builder import add_documentation, set_schema_options
 from openapi_builder.documentation import DiscriminatorOptions
+
+
+class Fish(halogen.Schema):
+    """Fish class."""
+
+    name = halogen.Attr(halogen.types.String())
+    """Name of this fish."""
 
 
 @pytest.fixture
@@ -16,10 +25,19 @@ def documentation_options__include_marshmallow_converters():
     return False
 
 
-@pytest.mark.usefixtures("get_with_halogen_schema")
-def test_get_halogen_string_schema(
-    http, halogen_schema, second_halogen_schema, open_api_documentation
-):
+def test_get_halogen_string_schema(http, app, open_api_documentation):
+    halogen_schema = halogen.Schema(
+        field=halogen.Attr(halogen.types.String()),
+    )
+    second_halogen_schema = halogen.Schema(
+        field=halogen.Attr(halogen.types.Int()),
+    )
+
+    @app.route("/get")
+    @add_documentation(response=halogen_schema)
+    def get():
+        return jsonify(halogen_schema.serialize({"field": "value"}))
+
     set_schema_options(
         schema=halogen_schema,
         discriminator=DiscriminatorOptions(
@@ -28,12 +46,12 @@ def test_get_halogen_string_schema(
             mapping={"second": second_halogen_schema},
         ),
     )
-    response = http.get("/get_with_halogen_schema")
+    response = http.get("/get")
     assert response.status_code == HTTPStatus.OK
     assert response.parsed_data == {"field": "value"}
 
     configuration = open_api_documentation.get_specification()
-    path = configuration["paths"]["/get_with_halogen_schema"]
+    path = configuration["paths"]["/get"]
     operation = path["get"]["responses"]["200"]
     assert operation["description"] == ""
     ref = operation["content"]["application/json"]["schema"]